% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adapter_content}
\alias{adapter_content}
\title{Get adapter content}
\usage{
adapter_content(infile, adapters)
}
\arguments{
\item{infile}{filepath to fastq sequence}

\item{adapters}{filepath to adapters}
}
\description{
Get adapter content
}
