% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_stat.R
\name{basic_stat}
\alias{basic_stat}
\title{Generate the data frame that includes percentiles of quality score per position}
\usage{
basic_stat(infile, writefile = FALSE, prefix)
}
\arguments{
\item{infile}{the object that is the dataframe of the mean, median and quantiles of the FASTQ file from basic statistics function}

\item{writefile}{the boolean of whether to write output as file}

\item{prefix}{the string of customized prefix of write-out file}
}
\value{
table of quality score statistics
}
\description{
Generate the data frame that includes percentiles of quality score per position
}
\examples{

basic_stat(system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq"),FALSE)

}
\author{
Wenyue Xing, \email{wenyue_xing@brown.edu}
}
