% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_adapter_content}
\alias{calc_adapter_content}
\title{Compute adapter content in reads.}
\usage{
calc_adapter_content(infile, adapters)
}
\arguments{
\item{infile}{filepath to fastq sequence}

\item{adapters}{filepath to adapters}
}
\value{
map object with adapter names as the key and the number of times the adapters appears in the reads as the value
}
\description{
Compute adapter content in reads.
}
