% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_format_score}
\alias{calc_format_score}
\title{Calculate score based on Illumina format}
\usage{
calc_format_score(score, score_format)
}
\arguments{
\item{score}{An ascii quality score from the fastq}

\item{score_format}{The illumina format}
}
\value{
a string as with the best guess as to the illumina format
}
\description{
Calculate score based on Illumina format
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
find_format(infile,10000, 100000)
}
