% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_over_rep_seq}
\alias{calc_over_rep_seq}
\title{Calculate sequece counts for each unique sequence and create a table with unique sequences
and corresponding counts}
\usage{
calc_over_rep_seq(infile, min_size = 5L, buffer_size = 1000000L)
}
\arguments{
\item{infile}{A string giving the path for the fastqfile}

\item{min_size}{An int for thhresholding over representation}

\item{buffer_size}{An int for the number of lines to keep in memory}
}
\value{
calculate overrepresented sequence count
}
\description{
Calculate sequece counts for each unique sequence and create a table with unique sequences
and corresponding counts
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
calc_over_rep_seq(infile)[seq_len(5)]
}
