% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_format}
\alias{find_format}
\title{Get Illumina format}
\usage{
find_format(infile, buffer_size, reads_used)
}
\arguments{
\item{infile}{A string giving the path for the fastqfile}

\item{buffer_size}{An int for the number of lines to keep in memory}
}
\value{
process fastq and evaluate the format of reads
}
\description{
Get Illumina format
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
find_format(infile,10000)
}
