% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_content.R
\name{GC_content}
\alias{GC_content}
\title{Calculates GC content percentage for each read in the dataset.}
\usage{
GC_content(infile, output_file = NA)
}
\arguments{
\item{infile}{the object that is the path to the FASTQ file}

\item{output_file}{File to write results to. Default NA.}
}
\value{
plot of GC content
}
\description{
Calculates GC content percentage for each read in the dataset.
}
\examples{

infile <- system.file("extdata", "10^5_reads_test.fq.gz",
    package = "qckitfastq")
head(GC_content(infile))

}
