% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gc_per_read}
\alias{gc_per_read}
\title{Calculate GC nucleotide sequence content per read of the FASTQ gzipped file}
\usage{
gc_per_read(infile)
}
\arguments{
\item{infile}{A string giving the path for the fastqfile}
}
\value{
GC content perncentage per read
}
\description{
Calculate GC nucleotide sequence content per read of the FASTQ gzipped file
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
gc_per_read(infile)
}
