% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmer_count.R
\name{Kmer_count}
\alias{Kmer_count}
\title{Return kmer count per sequence for the length of kmer desired}
\usage{
Kmer_count(infile, k, writefile = FALSE, prefix)
}
\arguments{
\item{infile}{the object that is the path to gzippped FASTQ file}

\item{k}{the length of kmer}

\item{writefile}{TRUE for writing the output to csv file}

\item{prefix}{the prefix of the output file if writefile is TRUE}
}
\value{
Kmers counts per sequence
}
\description{
Return kmer count per sequence for the length of kmer desired
}
