% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrep_sequence.R
\name{overrep_sequence}
\alias{overrep_sequence}
\title{Sort all sequences per read by count.}
\usage{
overrep_sequence(infile, nr, output_file = NA)
}
\arguments{
\item{infile}{the object that is the path to gzippped FASTQ file}

\item{nr}{the number of reads of the FASTQ file}

\item{output_file}{File to save data frame to. Default NA.}
}
\value{
Table of sequences sorted by count.
}
\description{
Sort all sequences per read by count.
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz",
    package = "qckitfastq")
overrep_sequence(infile,nr=25000)[1:5]
}
