% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_read_quality.R
\name{per_read_quality}
\alias{per_read_quality}
\title{Compute the mean quality score per read.
\code{per_read_quality}}
\usage{
per_read_quality(infile, output_file = NA)
}
\arguments{
\item{infile}{Path to FASTQ file}

\item{output_file}{File to write plot to. Will not write to file if NA. Default NA.}
}
\value{
Data frame of mean quality score per read
}
\description{
Compute the mean quality score per read.
\code{per_read_quality}
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
prq <- per_read_quality(infile)

}
