% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perseq_quality.R
\name{perseq_quality}
\alias{perseq_quality}
\title{Compute the mean quality score per sequence.
\code{perseq_quality}}
\usage{
perseq_quality(infile, output_file = NA)
}
\arguments{
\item{infile}{Path to FASTQ file}

\item{output_file}{File to write plot to. Will not write to file if NA. Default NA.}
}
\value{
Data frame of mean quality score per read
}
\description{
Compute the mean quality score per sequence.
\code{perseq_quality}
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
perseq_quality <- perseq_quality(infile)
}
