% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overrep_kmer.R
\name{plot_outliers}
\alias{plot_outliers}
\title{Determine how to plot outliers. Heuristic used is whether their obsexp_ratio differs by more than 1
and whether they fall into the same bin or not. If for 2 outliers, obsexp_ratio differs by less than .4
and they are in the same bin, then combine into a single plotting point.
NOT FULLY FUNCTIONAL}
\usage{
plot_outliers(overkm, top_num)
}
\arguments{
\item{overkm}{data frame with columns pos, obsexp_ratio, and kmer that has already been reordered by descending obsexp_ratio}

\item{top_num}{number of most overrepresented kmers to plot. Default is 5.}
}
\description{
Determine how to plot outliers. Heuristic used is whether their obsexp_ratio differs by more than 1
and whether they fall into the same bin or not. If for 2 outliers, obsexp_ratio differs by less than .4
and they are in the same bin, then combine into a single plotting point.
NOT FULLY FUNCTIONAL
}
