% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overrep_kmer.R
\name{plot_overrep_kmer}
\alias{plot_overrep_kmer}
\title{Create a box plot of the log2(observed/expected) ratio across the length of the sequence as well as top
overrepresented kmers. Only ratios greater than 2 are included in the box plot. Default is 20 bins across
the length of the sequence and the top 2 overrepresented kmers, but this can be changed by the user.}
\usage{
plot_overrep_kmer(overkm, bins = 20, top_num = 2, output_file = NA)
}
\arguments{
\item{overkm}{data frame with columns pos, obsexp_ratio, and kmer}

\item{bins}{number of intervals across the length of the sequence}

\item{top_num}{number of most overrepresented kmers to plot}

\item{output_file}{File to write plot to. Will not write to file if NA. Default NA.}
}
\value{
A box plot of the log2(observed/expected ratio) across the length of the sequence
}
\description{
Create a box plot of the log2(observed/expected) ratio across the length of the sequence as well as top
overrepresented kmers. Only ratios greater than 2 are included in the box plot. Default is 20 bins across
the length of the sequence and the top 2 overrepresented kmers, but this can be changed by the user.
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz",
    package = "qckitfastq")
over_km <- overrep_kmer(infile,k=4)
plot_overrep_kmer(over_km)
}
