% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_per_base_quality.R
\name{plot_per_base_quality}
\alias{plot_per_base_quality}
\title{Generate a boxplot of the per position quality score.}
\usage{
plot_per_base_quality(per_base_quality, output_file = NA)
}
\arguments{
\item{per_base_quality}{a data frame of the mean, median and quantiles of sequence quality per base. Most likely generated with the `per_base_quality` function.}

\item{output_file}{File to save plot to. Will not write to file if NA. Default NA.}
}
\value{
A boxplot of per position quality score distribution.
To address R CMD check from giving a NOTE about undefined global variables
https://dplyr.tidyverse.org/articles/programming.html
}
\description{
Generate a boxplot of the per position quality score.
}
\examples{

pbq <- per_base_quality(system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq"))
plot_per_base_quality(pbq)
}
