% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_per_read_quality.R
\name{plot_per_read_quality}
\alias{plot_per_read_quality}
\title{Plot the mean quality score per sequence as a histogram.
High quality sequences are those mostly distributed over 30.
Low quality sequences are those mostly under 30.
\code{plot_per_read_quality}}
\usage{
plot_per_read_quality(prq, output_file = NA)
}
\arguments{
\item{prq}{Data frame from per_read_quality function}

\item{output_file}{File to write plot to. Will not write to file if NA. Default NA.}
}
\value{
Plot of mean quality score per read
}
\description{
Plot the mean quality score per sequence as a histogram.
High quality sequences are those mostly distributed over 30.
Low quality sequences are those mostly under 30.
\code{plot_per_read_quality}
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
prq <- per_read_quality(infile)
plot_per_read_quality(prq)
}
