% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_perseq_quality.R
\name{plot_perseq_quality}
\alias{plot_perseq_quality}
\title{Plot the mean quality score per sequence as a histogram.
High quality sequences are those mostly distributed over 30.
Low quality sequences are those mostly under 30.
\code{plot_perseq_quality}}
\usage{
plot_perseq_quality(infile, output_file = NA)
}
\arguments{
\item{infile}{Path to FASTQ file}

\item{output_file}{File to write plot to. Will not write to file if NA. Default NA.}
}
\value{
Plot of mean quality score per read
}
\description{
Plot the mean quality score per sequence as a histogram.
High quality sequences are those mostly distributed over 30.
Low quality sequences are those mostly under 30.
\code{plot_perseq_quality}
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
plot_perseq_quality(infile)
}
