% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_content.R
\name{plot_read_content}
\alias{plot_read_content}
\title{Plot the per position nucleotide content.}
\usage{
plot_read_content(read_content, output_file = NA)
}
\arguments{
\item{read_content}{Data frame produced by read_content function.}

\item{output_file}{File to save plot to. Will not write to file if NA. Default NA.}
}
\value{
ggplot line plot of all nucleotide content inclding  A, T, G, C and N
}
\description{
Plot the per position nucleotide content.
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
fseq <- seqTools::fastqq(infile,k=6)
read_content <- read_content(fseq)
plot_read_content(read_content)
}
