% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_length.R
\name{plot_read_length}
\alias{plot_read_length}
\title{Extract the read length per read and plot a corresponding histogram of the number
of reads with each read length.}
\usage{
plot_read_length(fseq, output_file = NA)
}
\arguments{
\item{fseq}{a seqTools object produced by seqTools:fastqq on the raw FASTQ file.}

\item{output_file}{If specified, the output file to write to. Default is NA, i.e. do not write to file.}
}
\value{
A histogram of the read length distribution.
}
\description{
Extract the read length per read and plot a corresponding histogram of the number
of reads with each read length.
}
\examples{

infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
fseq <- seqTools::fastqq(infile,k=6)
plot_sequence_length(fseq)

}
\author{
Wenyue Xing, \email{wenyue_xing@brown.edu}, August Guang, \email{august_guang@brown.edu}
}
