% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sequence_length.R
\name{plot_sequence_length}
\alias{plot_sequence_length}
\title{extract the sequence length per read and plot corresponding bar plot}
\usage{
plot_sequence_length(fseq, writefile = FALSE, prefix)
}
\arguments{
\item{fseq}{the object that is the seqTools processed result}

\item{writefile}{the boolean object to write the plot as PDF file, default is FALSE}

\item{prefix}{the prefix to add to the file name}
}
\value{
the plot of the sequence distribution among all reads
}
\description{
extract the sequence length per read and plot corresponding bar plot
}
\examples{

infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
fseq <- seqTools::fastqq(infile,k=6)
plot_sequence_length(fseq)
}
\author{
Wenyue Xing, \email{wenyue_xing@brown.edu}
}
