% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_content.R
\name{process_adapter_file}
\alias{process_adapter_file}
\title{Helper function to process adapter file.}
\usage{
process_adapter_file(adapter_file)
}
\arguments{
\item{adapter_file}{Text file storing adapters names and sequences.}
}
\value{
data table
}
\description{
Helper function to process adapter file.
}
