% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{process_fastq}
\alias{process_fastq}
\title{Process fastq and generate sequence and quality score tables}
\usage{
process_fastq(infile, buffer_size)
}
\arguments{
\item{infile}{A string giving the path for the fastqfile}

\item{buffer_size}{An int for the number of lines to keep in memory}
}
\value{
process fastq and generate sequence and quality score tables
}
\description{
Process fastq and generate sequence and quality score tables
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
process_fastq(infile,10000)
}
