% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_content.R
\name{read_base_content}
\alias{read_base_content}
\title{Compute nucleotide content per position for a single base pair. Wrapper function around seqTools.}
\usage{
read_base_content(fseq, content)
}
\arguments{
\item{fseq}{a seqTools::fastqq object}

\item{content}{nucleotide. Options are "A","T","G","C","N"(either capital or lower case)}
}
\value{
Nucleotide sequence content per position.
}
\description{
Compute nucleotide content per position for a single base pair. Wrapper function around seqTools.
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
fseq <- seqTools::fastqq(infile,k=6)
read_base_content(fseq,"A")
}
\author{
Wenyue Xing, \email{wenyue_xing@brown.edu}, August Guang \email{august_guang@brown.edu}
}
