% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_content.R
\name{read_content}
\alias{read_content}
\title{Compute nucleotide content per position. Wrapper function around seqTools.}
\usage{
read_content(fseq, output_file = NA)
}
\arguments{
\item{fseq}{a seqTools::fastqq object}

\item{output_file}{File to write results in CSV format to. Will not write to file if NA. Default NA.}
}
\value{
Data frame of nucleotide sequence content per position
}
\description{
Compute nucleotide content per position. Wrapper function around seqTools.
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
fseq <- seqTools::fastqq(infile,k=6)
read_content(fseq)
}
