% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_length.R
\name{read_length}
\alias{read_length}
\title{Creates a data frame of read lengths and the number of reads with
that read length.}
\usage{
read_length(fseq, output_file = NA)
}
\arguments{
\item{fseq}{a seqTools object produced by seqTools::fastqq on the raw FASTQ file}

\item{output_file}{File to save data frame to. Default NA.}
}
\value{
Data frame of read lengths and number of reads with that length.
}
\description{
Creates a data frame of read lengths and the number of reads with
that read length.
}
\examples{
infile <- system.file("extdata","test.fq.gz",
    package = "qckitfastq")
fseq <- seqTools::fastqq(infile,k=6)
read_len <- read_length(fseq)
}
