% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_content.R
\name{sequence_content}
\alias{sequence_content}
\title{Compute nucleoctide sequence content per position. Wrapper function around seqTools.}
\usage{
sequence_content(fseq, content, output_file = NA)
}
\arguments{
\item{fseq}{a seqTools::fastqq object}

\item{content}{an object of string type that specifies the content in question, "A","T","G","C","N"(either capital or lower case)}

\item{output_file}{File to write results in CSV format to. Will not write to file if NA. Default NA.}
}
\value{
Nucleotide sequence content per position.
}
\description{
Compute nucleoctide sequence content per position. Wrapper function around seqTools.
}
\examples{
infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
fseq <- seqTools::fastqq(infile,k=6)
sequence_content(fseq,"A")
}
\author{
Wenyue Xing, \email{wenyue_xing@brown.edu}, August Guang \email{august_guang@brown.edu}
}
