% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/adapter_content.R
\name{adapter_content}
\alias{adapter_content}
\title{Get adapter content}
\usage{
adapter_content(infile, adapter_file)

adapter_content(infile, adapter_file)
}
\arguments{
\item{infile}{filepath to fastq sequence}

\item{adapter_file}{filepath to the adapter sequences to search for. Default is extdata/adapters.txt.}

\item{adapters}{filepath to adapters}

\item{infile}{the path to the FASTQ file}
}
\value{
proportion of adapter sequences]
}
\description{
Get adapter content

Search through sequences and return the total proportion of the library which contains
adapter sequences, predefined in the file adapters.txt. It is also possible for the user to
give a different file as input; the function is relatively generic. Sequences however must
be in the same format as adapters.txt, that is name tab sequence.
}
