% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapter_content.R
\name{adapter_content}
\alias{adapter_content}
\title{Search through sequences and return the total proportion of the library which contains
adapter sequences, predefined in the file adapters.txt. It is also possible for the user to
give a different file as input; the function is relatively generic. Sequences however must
be in the same format as adapters.txt, that is name tab sequence.}
\usage{
adapter_content(infile, adapter_file)
}
\arguments{
\item{infile}{the path to the FASTQ file}

\item{adapter_file}{filepath to the adapter sequences to search for. Default is extdata/adapters.txt.}
}
\value{
proportion of adapter sequences
}
\description{
Search through sequences and return the total proportion of the library which contains
adapter sequences, predefined in the file adapters.txt. It is also possible for the user to
give a different file as input; the function is relatively generic. Sequences however must
be in the same format as adapters.txt, that is name tab sequence.
}
