% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{dimensions}
\alias{dimensions}
\title{Extract the number of columns and rows for a FASTQ file using seqTools.}
\usage{
dimensions(fseq, selection)
}
\arguments{
\item{fseq}{an object that is the read result of the seq.read function}

\item{selection}{"reads' for number of reads/rows, 'positions' for number of positions/columns}
}
\value{
a numeric value of the number of reads or the number of positions
}
\description{
Extract the number of columns and rows for a FASTQ file using seqTools.
}
\examples{

infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
fseq <- seqTools::fastqq(infile,k=6)
dimensions(fseq,"reads")

}
