% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GC_content.R
\name{GC_content}
\alias{GC_content}
\title{Extract GC content separately and calculate GC content percentage for each sequence read}
\usage{
GC_content(infile, writefile = FALSE, prefix)
}
\arguments{
\item{infile}{the object that is the path to the FASTQ file}

\item{writefile}{TRUE for writing the output to csv file}

\item{prefix}{the prefix of the output file if writefile is TRUE}
}
\value{
plot of GC content
}
\description{
Extract GC content separately and calculate GC content percentage for each sequence read
}
\examples{

infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
GC_content(infile,FALSE)
}
