% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overrep_sequence.R
\name{overrep_sequence}
\alias{overrep_sequence}
\title{Sort all sequences per read by count along with a density plot of all counts with top 5 repreated sequences marked}
\usage{
overrep_sequence(infile, nr, writefile = FALSE, prefix)
}
\arguments{
\item{infile}{the object that is the path to gzippped FASTQ file}

\item{nr}{the number of reads of the FASTQ file}

\item{writefile}{TRUE for writing the output to csv file}

\item{prefix}{prefix of output file if writefile is TRUE}
}
\value{
table of sequnces sortted by count

density plot of sequence length with top 5 marked by rugs, saved as PDF file
}
\description{
Sort all sequences per read by count along with a density plot of all counts with top 5 repreated sequences marked
}
