% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_overep_seq.R
\name{plot_overrep_seq}
\alias{plot_overrep_seq}
\title{Plot the top 5 seqeunces}
\usage{
plot_overrep_seq(overrep_order, output_file = NA)
}
\arguments{
\item{overrep_order}{the table that sorts the sequence content and corresponding counts in descending order}

\item{output_file}{File to save plot to. Will not write to file if NA. Default NA.}
}
\value{
plot of the top 5 overrepresented sequences
}
\description{
Plot the top 5 seqeunces
}
\examples{

infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
overrep_order <- overrep_sequence(infile,nr=25000,"test")
plot_overrep_seq(overrep_order)

}
