% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_read_length.R
\name{plot_read_length}
\alias{plot_read_length}
\title{Plot a histogram of the number of reads with each read length.}
\usage{
plot_read_length(read_len, output_file = NA)
}
\arguments{
\item{read_len}{Data frame of read lengths and number of reads with that length.}

\item{output_file}{File to save plot to. Default is NA, i.e. do not write to file.}
}
\value{
A histogram of the read length distribution.
}
\description{
Plot a histogram of the number of reads with each read length.
}
\examples{

infile <- system.file("extdata", "10^5_reads_test.fq.gz", package = "qckitfastq")
fseq <- seqTools::fastqq(infile,k=6)
read_len <- read_length(fseq)
plot_read_length(read_len)

}
\author{
Wenyue Xing, \email{wenyue_xing@brown.edu}, August Guang, \email{august_guang@brown.edu}
}
