% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{One function to run them all}
\usage{
run_all(infile, foption, poption, pref)
}
\arguments{
\item{infile}{the path to the datafile, can be fastq file either g-zipped or not}

\item{foption}{TRUE for saving all output results to file, FALSE for explicitly outputting the output tables}

\item{poption}{TRUE for saving all plots to file, FALSE for displaying all plots}

\item{pref}{if any above choice is TRUE, give the prefix to the customized output filename prefix}
}
\value{
generate files from all functions
}
\description{
One function to run them all
}
